move process_queue to queue module
